bpf:cgroup: add bpf_cgroup_attach_task bpf program
This bpf program will run under "raw_tracepoint/cgroup_attack_task" that is
triggered when a task is migrated to a cgroup.

We use this as a detection point for cgroup configurations. When tetragon
userspace starts it will write its own NSPID (pid if in init pid namespace)
to its own same cgroup, this allows to discover tetragon information
(pid, assert pid, cgroup level, cgroup id...)

The detection can be improved in future.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>