bpf:cgroup: set the tracking Cgroup ID of tasks during fork
When a task forks we will query the cgroup tracking level, if it has
a deeper nested level then we use the Tetragon cgroup level and assign it to
the task execve_map_value in the execve_map. This way all forked childs
will belong to the same tracked cgroup.

If it has a lower cgroup tracking level, this is probably a service, pod,
container being created and in this case we assign the current task cgroup
id as the tracking cgroup id.

This will allow later Tetragon to track all processes that have started
inside a container/pod easily, where they will be assigned a cgroup tracking
ID that should not change during their life cycle even if the task's cgroup
config changes, migrates within pod, etc. the tracking is pinned for the
task life time.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>