bpf:cgroups: pass subsys index when operating on cgroup_subsys_state set
Specify the subsystem index when operating on cgroup_subsys_state set.

This allows to support multiple setups where cgroup environment may not
be configured in a standard way including cgroupv1 and v2, or if the
hierarchy is missing some controllers. However we require that the 'memory'
or 'pids' controllers to be set which should always be the case. If they
are not properly setup then Tetragon cgroup tracking will not operate correctly
and probably we should not support such workload.

The bpf helpers will read the subsystem index from Tetragon Conf that
was set dynamically by user space during startup, it should point on
one of the 'memory' or 'pids' controllers.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>