bpf:cgroup: improve Cgroup detection by checking Cgroup ID and hierarchy
Improve cgroup attach detection by checking Cgroup ID and current Hierarchy
ID if both match. This will ensure that current task is really operating
on the right Cgroup ID that belongs to the right Cgroup hierarchy. Using
namespaced PID may clash if we have another process with same pid in
different namespace PID that is doing same thing... anyway with this patch
this is robust now and impossible to clash.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>