bpf:cgroup: set the tracking Cgroup ID of tasks during fork
When a task forks we will query the cgroup tracking level, if it has
a deeper nested level then we use the Tetragon cgroup level mapping it
to its current cgroup ancestor that is at the same level and assign it to
the task execve_map_value in the execve_map. This way all forked childs
will belong to the same tracked cgroup.

If it has a lower cgroup tracking level, then it is ancestor, this is
probably a service, pod or container being created and in this case we
use the current cgroup id as the tracking cgroup id.

This will allow later Tetragon to track all processes that have started
inside a container/pod easily, where they will be assigned a cgroup tracking
ID when forked that should not change during their life cycle even if the
task's cgroup config changes, migrates within pod, etc.

The tracking is pinned for the task life time.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>