pkg:cgroups: get Tetragon own Cgroup ID and pass it to BPF
Get Tetragon own Cgroup ID and pass it to BPF so we can improve
Tetragon own detection and assert that this is Tetragon user space
trying to probe cgroup detection.

Also this can be useful in future in order to allow Tetragon user
space do some operations without filtering ourselves.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>